New improved way of communication with Api - network errors like host not available are now being centrally handled.

+++ app/src/main/java/cz/mendelu/xmusil5/plantmonitor/navigation/INavigationRouter.kt
@@ -13,8 +13,8 @@ interface INavigationRouter {
     fun toDevicesScreen()
     fun toProfileScreen()
 
-    fun toPlantDetail(plantId: Int)
-    fun toDeviceDetail(deviceId: Int)
+    fun toPlantDetail(plantId: Long)
+    fun toDeviceDetail(deviceId: Long)
 
     fun toAddPlant()
     fun toAddDevice()
